Moving and fixing code
Let me know if you have anymore issues or concerns.
